requirements.txt
setuptools~=49.3.1.post20200810
PyYAML~=5.1.2
kubernetes~=11.0.0
requests~=2.18.4